requirements.txt
# 🔥 ULTIMATE BONZAI BACKEND REQUIREMENTS
# Compiled from backend-1, backend-2, backend-3, and backend-6
# Supports: Virtual Computer, WhatsApp, DeepSeek, CrewAI, Monitoring, Multi-Provider

# ===== CORE FLASK ECOSYSTEM =====
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.6
flask-limiter==3.5.0
python-socketio==5.8.0
python-socketio[client]==5.8.0
eventlet==0.33.3

# ===== WEB SERVERS & ASYNC =====
uvicorn==0.24.0
gunicorn==21.2.0
aiohttp==3.9.1
aiofiles==23.2.1
websockets==12.0
psutil==5.9.6

# ===== AI MODEL PROVIDERS =====
# Anthropic (Claude)
anthropic==0.34.2

# OpenAI (GPT)
openai==1.6.1

# Google (Gemini)
google-generativeai==0.7.2
google-cloud-aiplatform==1.40.0
google-auth==2.25.2

# Memory & Vector DB
mem0ai==0.0.11
qdrant-client==1.6.9

# ===== WEB SCRAPING & HTTP =====
requests==2.31.0
beautifulsoup4==4.12.2
scrapybara==1.0.0

# ===== DATABASE & CACHING =====
redis==5.0.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# ===== DATA PROCESSING & ML =====
numpy==1.24.4
scipy==1.11.4
scikit-learn==1.3.2
pydantic==2.5.3
dataclasses-json==0.6.0

# ===== CODE EXECUTION & INTERPRETATION =====
e2b-code-interpreter==1.5.1

# ===== SYSTEM & UTILITIES =====
psutil==5.9.6
python-dotenv==1.0.0
colorama==0.4.6
rich==13.7.0
python-dateutil==2.8.2

# ===== ASYNC & CONCURRENCY =====
asyncio-throttle==1.0.2

# ===== WHATSAPP INTEGRATION =====
# WhatsApp dependencies (may need additional setup)
twilio>=8.0.0
qrcode>=7.4.2
pillow>=10.0.0

# ===== CREWAI & AGENT ORCHESTRATION =====
crewai>=0.1.0
langchain>=0.1.0
langchain-community>=0.0.1

# ===== DEEPSEEK INTEGRATION =====
# DeepSeek SDK and related dependencies
httpx>=0.25.0
tenacity>=8.2.0

# ===== MONITORING & OBSERVABILITY =====
prometheus-client>=0.18.0
logging>=0.4.9.6

# ===== DEVELOPMENT & TESTING =====
pytest>=7.4.0
pytest-asyncio>=0.21.0

# ===== UTILITIES & PARSING =====
uuid==1.30
json5==0.9.14
pathlib==1.0.1

# ===== VIRTUAL COMPUTER DEPENDENCIES =====
# For virtual computer service integration
docker>=6.1.0
paramiko>=3.3.0

# ===== MCP (Model Context Protocol) =====
# For MCP client/server integration
mcp>=0.1.0

# ===== IDE INTEGRATION =====
# For agentic IDE integration
jupyterlab>=4.0.0
notebook>=7.0.0

# ===== ENHANCED FEATURES =====
# Additional packages for advanced functionality
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.1.0

# Note: Some packages may require additional system dependencies
# WhatsApp integration may need Chrome/Chromium for web automation
# Virtual computer service may need Docker runtime
# IDE integration may need Jupyter system packages